+++ src/main/java/ModuloOperator.java
@@ -1,11 +1,11 @@
 public class ModuloOperator extends Operator {
 
   @Override
-  public int apply(int operand1, int operand2) throws DummyFloatingPointException {
+  public double apply(double operand1, double operand2) throws DummyFloatingPointException {
     if (operand2 == 0) {
       throw new DummyFloatingPointException();
     }
-    return operand1 % operand2;
+    return operand1 % operand2; // TODO this can only be applied with an integer
   }
 
   @Override
