+++ src/test/java/SingleCharacterTokenTokenizerTest.java
@@ -76,7 +76,7 @@ public class SingleCharacterTokenTokenizerTest {
     }
 
     @Override
-    public void apply(SizeRestrictedStack<Integer> operandStack) throws CalculatorException {}
+    public void apply(SizeRestrictedStack<Double> operandStack) throws CalculatorException {}
 
     public char getValue() {
       return value;
